+ Entire authentication overhaul
